Windows, web-ui refactoring continued, removed mounting node_modules on everything.
Made it work on windows.
Still going through web-ui to fix the calls. Changed file uploading to just use multer. Apparently some foundation stuff wasn't there or something. That was weird.
Need to do the same change to the other services, and do something about the original file...or just put it in memory.